requirements.txt
Flask==2.3.3
Pillow==10.0.0
requests==2.31.0
APScheduler==3.10.4
python-dotenv==1.0.0
arabic-reshaper==3.0.0
python-bidi==0.4.2
Werkzeug==2.3.7
gunicorn==21.2.0
click==8.1.7
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.3
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
tzlocal==5.0.1
# Added for font management
google-fonts-downloader>=0.1.0
# Added for testing
pytest==7.4.0 